+++ rector.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Config\RectorConfig;
+use Rector\Set\ValueObject\LevelSetList;
+use Rector\Set\ValueObject\SetList;
+
+return RectorConfig::configure()
+    ->withPaths([
+        __DIR__ . '/src',
+        __DIR__ . '/config',
+    ])
+    ->withPhpSets()
+    ->withTypeCoverageLevel(0)
+    ->withSets([
+        SetList::DEAD_CODE,
+        SetList::EARLY_RETURN,
+        SetList::CODE_QUALITY,
+    ])
+    ->withImportNames(removeUnusedImports: true)
+    ->withPreparedSets(
+        deadCode: true,
+        codeQuality: true,
+        earlyReturn: true,
+    );